+++ course_learning/src/models/associations/slide-template-admin.associations.ts
