+++ course_learning/src/models/associations/slide.associations.ts
@@ -1,6 +1,6 @@
 import { Level } from '../level';
 import { Slide } from '../slide';
-import { SlideTemplate } from '../slideTemplate';
+import { SlideTemplate } from '../slide-template';
 
 const defineSlideAssociations = () => {
   Slide.belongsTo(Level, { foreignKey: 'levelId' });
