+++ course_learning/src/models/slide-template.ts
