+++ course_learning/src/routes/show-hints.ts
